darkiworld/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
undetected-chromedriver==3.5.5
selenium==4.16.0
